# Optimize WebGPU ArgMinMax reduction factor